requirements.txt
Pillow==9.2.0
python_ev3dev2==2.1.0.post1
Telethon==1.25.0
